refactor : move member domain
