+++ backend/src/main/java/toonpick/app/domain/bookmark/BookmarkToons.java
@@ -15,7 +15,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import toonpick.app.common.utils.BaseTimeEntity;
-import toonpick.app.member.entity.Member;
+import toonpick.app.domain.member.Member;
 import toonpick.app.domain.webtoon.Webtoon;
 
 @Entity
