+++ backend/src/main/java/toonpick/app/domain/bookmark/BookmarkToonsService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.exception.ResourceNotFoundException;
-import toonpick.app.member.entity.Member;
-import toonpick.app.member.repository.MemberRepository;
+import toonpick.app.domain.member.Member;
+import toonpick.app.repository.MemberRepository;
 import toonpick.app.domain.webtoon.Webtoon;
 import toonpick.app.repository.WebtoonRepository;
