+++ backend/src/main/java/toonpick/app/domain/collection/ToonCollectionRepository.java
@@ -2,7 +2,7 @@ package toonpick.app.domain.collection;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import toonpick.app.member.entity.Member;
+import toonpick.app.domain.member.Member;
 
 import java.util.List;
