+++ backend/src/main/java/toonpick/app/domain/member/MemberFavoriteWebtoon.java
@@ -1,4 +1,4 @@
-package toonpick.app.member.entity;
+package toonpick.app.domain.member;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
