+++ backend/src/main/java/toonpick/app/repository/MemberFavoriteWebtoonRepository.java
@@ -1,11 +1,11 @@
-package toonpick.app.member.repository;
+package toonpick.app.repository;
 
 import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import toonpick.app.member.entity.Member;
-import toonpick.app.member.entity.MemberFavoriteWebtoon;
+import toonpick.app.domain.member.Member;
+import toonpick.app.domain.member.MemberFavoriteWebtoon;
 import toonpick.app.domain.webtoon.Webtoon;
 
 import java.util.List;
