+++ backend/src/main/java/toonpick/app/review/dto/WebtoonReviewDTO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import toonpick.app.member.dto.MemberDTO;
+import toonpick.app.dto.MemberDTO;
 
 import java.time.LocalDateTime;
