+++ backend/src/main/java/toonpick/app/review/entity/ReviewLike.java
@@ -13,7 +13,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import toonpick.app.member.entity.Member;
+import toonpick.app.domain.member.Member;
 
 @Entity
 @Getter
