+++ backend/src/main/java/toonpick/app/security/service/JoinService.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import toonpick.app.security.dto.JoinRequest;
 import toonpick.app.security.exception.UsernameAlreadyExistsException;
-import toonpick.app.member.entity.Member;
-import toonpick.app.member.repository.MemberRepository;
+import toonpick.app.domain.member.Member;
+import toonpick.app.repository.MemberRepository;
 
 import java.util.Random;
