+++ backend/src/main/java/toonpick/app/security/user/CustomOAuth2UserDetails.java
@@ -2,7 +2,7 @@ package toonpick.app.security.user;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
-import toonpick.app.member.dto.MemberDTO;
+import toonpick.app.dto.MemberDTO;
 
 import java.util.ArrayList;
 import java.util.Collection;
