+++ backend/src/main/java/toonpick/app/security/user/CustomUserDetails.java
@@ -3,7 +3,7 @@ package toonpick.app.security.user;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import toonpick.app.member.entity.Member;
+import toonpick.app.domain.member.Member;
 
 import java.util.Collection;
 import java.util.Collections;
