+++ backend/src/test/java/toonpick/app/unit/service/JoinServiceTest.java
@@ -7,8 +7,8 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import toonpick.app.security.dto.JoinRequest;
-import toonpick.app.member.entity.Member;
-import toonpick.app.member.repository.MemberRepository;
+import toonpick.app.domain.member.Member;
+import toonpick.app.repository.MemberRepository;
 import toonpick.app.security.service.JoinService;
 
 import static org.mockito.Mockito.*;
